+++ core/src/test/java/com/syntaric/openfhir/mapping/kds/KdsToAqlTest.java
@@ -91,7 +91,7 @@ public class KdsToAqlTest {
 
         final HelpersCreator helpersCreator1 = new HelpersCreator(repo, new AqlToFlatPathConverter(
                 new OpenFhirStringUtils(),
-                new OpenFhirMapperUtils()), new OpenFhirStringUtils());
+                new OpenFhirMapperUtils()));
         toAql = new ToAql(fhirConnectManager, new OpenFhirMapperUtils(), repo, new ToAqlMappingEngine(new OpenEhrAqlPopulator()), helpersCreator1,
                 new OpenEhrTemplateUtils(), null, optManager);
     }